requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
Pillow==9.5.0
requests==2.31.0 